+++ hpcg/src-mixed/CheckAspectRatio.hpp
@@ -14,6 +14,6 @@
 
 #ifndef CHECKASPECTRATIO_HPP
 #define CHECKASPECTRATIO_HPP
-extern int CheckAspectRatio(double smallest_ratio, int x, int y, int z, const char *what, bool DoIo);
+extern int CheckAspectRatio(float smallest_ratio, int x, int y, int z, const char *what, bool DoIo);
 #endif // CHECKASPECTRATIO_HPP
