+++ hpcg/src-mixed/ComputeOptimalShapeXYZ.cpp
@@ -9,7 +9,7 @@
 
 static void
 ComputePrimeFactors(int n, std::map<int, int> & factors) {
-  int d, sq = int((sqrt(double(n)))+1L);
+  int d, sq = int((sqrt(float(n)))+1L);
   div_t r;
 
   // remove 2 as a factor with shifts instead "/" and "%"
@@ -92,7 +92,7 @@ ComputeOptimalShapeXYZ(int xyz, int & x, int & y, int & z) {
     MixedBaseCounter c_main(count_factors, factors.size()), c1(count_factors, factors.size());
 
     // at the beginning, minimum area is the maximum area
-    double area, min_area = 2.0 * xyz + 1.0;
+    float area, min_area = 2.0 * xyz + 1.0;
 
     for (c1.next(); ! c1.is_zero(); c1.next()) {
       MixedBaseCounter c2(c_main, c1); // "c2" gets the factors remaining in "c_main" that "c1" doesn't have
@@ -101,7 +101,7 @@ ComputeOptimalShapeXYZ(int xyz, int & x, int & y, int & z) {
         int tf2 = c2.product(distinct_factors);
         int tf3 = xyz / tf1/ tf2; // we derive the third dimension, we don't keep track of the factors it has
 
-        area = tf1 * double(tf2) + tf2 * double(tf3) + tf1 * double(tf3);
+        area = tf1 * float(tf2) + tf2 * float(tf3) + tf1 * float(tf3);
         if (area < min_area) {
           min_area = area;
           x = tf1;
